+++ lib/Steps/ConfigureProject.ts
@@ -1,12 +1,9 @@
 import { Answers } from 'inquirer';
-import * as _ from 'lodash';
-import { ProjectType } from '../Constants';
-import { getCurrentProject } from '../Helper/Wizard';
+import { getCurrentIntegration } from '../Helper/Wizard';
 import { BaseStep } from './BaseStep';
-import { BaseProject } from './Projects/BaseProject';
 
 export class ConfigureProject extends BaseStep {
   public async emit(answers: Answers) {
-    return getCurrentProject(answers).emit(answers);
+    return getCurrentIntegration(answers).emit(answers);
   }
 }