+++ lib/Steps/Integrations/GenericNode.ts
@@ -1,11 +1,11 @@
 import { Answers } from 'inquirer';
 import * as _ from 'lodash';
 import { green, l, nl } from '../../Helper/Logging';
-import { BaseProject } from './BaseProject';
+import { BaseIntegration } from './BaseIntegration';
 
-export class GenericNode extends BaseProject {
+export class GenericNode extends BaseIntegration {
   public async emit(answers: Answers) {
-    const dsn = _.get(answers, 'selectedProject.keys.0.dsn.secret', null);
+    const dsn = _.get(answers, 'config.dsn.secret', null);
     if (!dsn) {
       return {};
     }