+++ lib/Steps/WaitForSentry.ts
@@ -2,14 +2,13 @@ import { Answers } from 'inquirer';
 import * as _ from 'lodash';
 import { BottomBar } from '../Helper/BottomBar';
 import { debug, dim, green, l, nl } from '../Helper/Logging';
-import { getCurrentProject } from '../Helper/Wizard';
+import { getCurrentIntegration } from '../Helper/Wizard';
 import { BaseStep } from './BaseStep';
-import { BaseProject } from './Projects/BaseProject';
 const r2 = require('r2');
 
 export class WaitForSentry extends BaseStep {
   public async emit(answers: Answers) {
-    if (!await getCurrentProject(answers).shouldEmit(answers)) {
+    if (!await getCurrentIntegration(answers).shouldEmit(answers)) {
       return {};
     }
     if (this.argv.skipConnect) {